requirements.txt
arrow==0.16.0
asgiref==3.3.0
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
binaryornot==0.4.4
certifi==2020.6.20
cookiecutter==1.7.2
decorator==4.4.2
Django==3.1.3
djangorestframework==3.12.2
gunicorn==20.0.4
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
jinja2-time==0.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
poyo==0.5.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycodestyle==2.6.0
Pygments==2.7.1
pylint==2.5.3
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
pyzmq==19.0.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
